+++ application/models/Providers_model.php
@@ -763,11 +763,26 @@ class Providers_model extends EA_Model
                 'googleSync' => array_key_exists('google_sync', $provider['settings'])
                     ? filter_var($provider['settings']['google_sync'], FILTER_VALIDATE_BOOLEAN)
                     : null,
+                'googleToken' => array_key_exists('google_token', $provider['settings'])
+                    ? $provider['settings']['google_token']
+                    : null,
                 'googleCalendar' => array_key_exists('google_calendar', $provider['settings'])
                     ? $provider['settings']['google_calendar']
                     : null,
-                'googleToken' => array_key_exists('google_token', $provider['settings'])
-                    ? $provider['settings']['google_token']
+                'caldavSync' => array_key_exists('caldav_sync', $provider['settings'])
+                    ? filter_var($provider['settings']['caldav_sync'], FILTER_VALIDATE_BOOLEAN)
+                    : null,
+                'caldavUrl' => array_key_exists('caldav_url', $provider['settings'])
+                    ? $provider['settings']['caldav_url']
+                    : null,
+                'caldavUsername' => array_key_exists('caldav_username', $provider['settings'])
+                    ? $provider['settings']['caldav_username']
+                    : null,
+                'caldavPassword' => array_key_exists('caldav_password', $provider['settings'])
+                    ? $provider['settings']['caldav_password']
+                    : null,
+                'caldavCalendar' => array_key_exists('caldav_calendar', $provider['settings'])
+                    ? $provider['settings']['caldav_calendar']
                     : null,
                 'syncFutureDays' => array_key_exists('sync_future_days', $provider['settings'])
                     ? (int) $provider['settings']['sync_future_days']